move login and logout management from site to admin
