+++ Yii2/controllers/AdminController.php
@@ -2,13 +2,14 @@
 
 namespace app\controllers;
 
-use Yii;
-use app\models\Comment;
+use app\models\LoginForm;
 use app\models\Post;
 use yii\data\ActiveDataProvider;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
+use yii\filters\AccessControl;
 use yii\filters\VerbFilter;
+use Yii;
 
 /**
  * AdminController implements the CRUD actions for Post model.
@@ -18,6 +19,17 @@ class AdminController extends Controller
     public function behaviors()
     {
         return [
+            'access' => [
+                'class' => AccessControl::className(),
+                'only' => ['logout'],
+                'rules' => [
+                    [
+                        'actions' => ['logout'],
+                        'allow' => true,
+                        'roles' => ['@'],
+                    ],
+                ],
+            ],
             'verbs' => [
                 'class' => VerbFilter::className(),
             ],
@@ -40,6 +52,49 @@ class AdminController extends Controller
         return $this->render('index', [
             'dataProvider' => $dataProvider,
         ]);
+    }
+
+    /**
+     * log the user in B.O
+     *
+     * @return null|string|\yii\web\Response
+     * @throws Exception
+     * @throws \Exception
+     */
+    public function actionLogin()
+    {
+        try {
+            Yii::trace('Trace : '.__METHOD__, __METHOD__);
+
+            $response = null;
+            $model = new LoginForm();
+            if ($model->load(Yii::$app->request->post()) && $model->login()) {
+                $response = $this->redirect(['/admin/index']);
+            }
+
+            if ($response === null) {
+                $response = $this->render('login', [
+                    'model' => $model,
+                ]);
+            }
+
+            return $response;
+        } catch(Exception $e) {
+            Yii::error($e->getMessage(), __METHOD__);
+            throw $e;
+        }
+    }
+
+    /**
+     * logout the user
+     *
+     * @return \yii\web\Response
+     */
+    public function actionLogout()
+    {
+        Yii::$app->user->logout();
+
+        return $this->goHome();
     }
 
     /**
